ajout semaphore pour jobs

+++ millegrilles_fichiers/Consignation.py
@@ -175,24 +175,33 @@ class ConsignationHandler:
         if self.__timestamp_visite is None or now - self.__intervalle_visites > self.__timestamp_visite:
             self.__logger.info("__traiter_cedule_local Visiter fuuids")
             try:
-                self.__timestamp_visite = datetime.datetime.utcnow()
-                await self.__store_consignation.visiter_fuuids()
+                # Demarrer la job si le semaphore n'est pas deja bloque
+                if self.__etat_instance.lock_db_job.locked() is False:
+                    async with self.__etat_instance.lock_db_job:
+                        self.__timestamp_visite = datetime.datetime.utcnow()
+                        await self.__store_consignation.visiter_fuuids()
             except Exception:
                 self.__logger.exception("__traiter_cedule_local Erreur visiter fuuids")
 
         if self.__timestamp_verification is None or now - self.__intervalle_verification > self.__timestamp_verification:
             self.__logger.info("__traiter_cedule_local Verifier fuuids")
             try:
-                self.__timestamp_verification = datetime.datetime.utcnow()
-                await self.__store_consignation.verifier_fuuids()
+                # Demarrer la job si le semaphore n'est pas deja bloque
+                if self.__etat_instance.lock_db_job.locked() is False:
+                    async with self.__etat_instance.lock_db_job:
+                        self.__timestamp_verification = datetime.datetime.utcnow()
+                        await self.__store_consignation.verifier_fuuids()
             except Exception:
                 self.__logger.exception("__traiter_cedule_local Erreur verifier fuuids")
 
         if self.__timestamp_orphelins is None or now - self.__intervalle_orphelins > self.__timestamp_orphelins:
             self.__logger.info("__traiter_cedule_local Supprimer orphelins")
             try:
-                self.__timestamp_orphelins = datetime.datetime.utcnow()
-                await self.__store_consignation.supprimer_orphelins()
+                # Demarrer la job si le semaphore n'est pas deja bloque
+                if self.__etat_instance.lock_db_job.locked() is False:
+                    async with self.__etat_instance.lock_db_job:
+                        self.__timestamp_orphelins = datetime.datetime.utcnow()
+                        await self.__store_consignation.supprimer_orphelins()
             except Exception:
                 self.__logger.exception("__traiter_cedule_local Erreur supprimer_orphelins")
